+++ CHANGELOG.md
@@ -26,6 +26,9 @@ We use the following categories for changes:
   Added `prom_info.metric_detail` which includes the disk size related columns. 
   Improved the performance of these. [#547]
 
+### Added
+- Telemetry for active series and last updated [#534]
+
 ## [0.7.0 - 2022-10-03]
 
 ### Added